Don't display Speaker + Deputy Speaker as parties
We should find a better way to deal with these, so they don’t end up
here in the first place, but for now simply ignore them explicitly.

Liquid has a ‘if string includes’, but I couldn’t find a ‘not
includes’, and the nested logic of an unless didn’t seem any better
than just listing both.